refactored app helpers into a module
